+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/admin/AdminApiDelayedDelivery.java
@@ -18,7 +18,6 @@
  */
 package org.apache.pulsar.broker.admin;
 
-import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import lombok.Cleanup;
 import lombok.extern.slf4j.Slf4j;
@@ -33,7 +32,6 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import java.util.HashSet;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.concurrent.TimeUnit;
@@ -48,8 +46,6 @@ public class AdminApiDelayedDelivery extends MockedPulsarServiceBaseTest {
     @BeforeMethod
     @Override
     public void setup() throws Exception {
-        conf.setSystemTopicEnabled(true);
-        conf.setTopicLevelPoliciesEnabled(true);
         super.internalSetup();
 
         admin.clusters().createCluster("test", new ClusterData(pulsar.getWebServiceAddress()));
@@ -74,122 +70,6 @@ public class AdminApiDelayedDelivery extends MockedPulsarServiceBaseTest {
 
         assertFalse(admin.namespaces().getDelayedDelivery(namespace).isActive());
         assertEquals(2000, admin.namespaces().getDelayedDelivery(namespace).getTickTime());
-    }
-
-    @Test(timeOut = 20000)
-    public void testEnableAndDisableTopicDelayedDelivery() throws Exception {
-        String topicName = "persistent://delayed-delivery-messages/default-ns/test" + System.currentTimeMillis();
-        String namespace = "delayed-delivery-messages/default-ns";
-        admin.namespaces().createNamespace(namespace);
-        admin.namespaces().setNamespaceReplicationClusters(namespace, Sets.newLinkedHashSet(Lists.newArrayList("test")));
-
-        admin.topics().createPartitionedTopic(topicName, 3);
-        assertNull(admin.topics().getDelayedDeliveryPolicy(topicName));
-
-        DelayedDeliveryPolicies delayedDeliveryPolicies = new DelayedDeliveryPolicies(2000, true);
-        admin.topics().setDelayedDeliveryPolicy(topicName, delayedDeliveryPolicies);
-        //wait for update
-        for (int i = 0; i < 50; i++) {
-            Thread.sleep(100);
-            if (admin.topics().getDelayedDeliveryPolicy(topicName) != null) {
-                break;
-            }
-        }
-
-        assertTrue(admin.topics().getDelayedDeliveryPolicy(topicName).isActive());
-        assertEquals(2000, admin.topics().getDelayedDeliveryPolicy(topicName).getTickTime());
-
-        admin.topics().removeDelayedDeliveryPolicy(topicName);
-        //wait for update
-        for (int i = 0; i < 50; i++) {
-            Thread.sleep(100);
-            if (admin.topics().getDelayedDeliveryPolicy(topicName) == null) {
-                break;
-            }
-        }
-        assertNull(admin.topics().getDelayedDeliveryPolicy(topicName));
-    }
-
-    @Test(timeOut = 20000)
-    public void testEnableTopicDelayedDelivery() throws Exception {
-        final String topicName = "persistent://delayed-delivery-messages/default-ns/test" + System.currentTimeMillis();
-        final String namespace = "delayed-delivery-messages/default-ns";
-        admin.namespaces().createNamespace(namespace);
-        admin.namespaces().setNamespaceReplicationClusters(namespace, Sets.newLinkedHashSet(Lists.newArrayList("test")));
-
-        admin.topics().createPartitionedTopic(topicName, 3);
-        assertNull(admin.topics().getDelayedDeliveryPolicy(topicName));
-        //1 Set topic policy
-        DelayedDeliveryPolicies delayedDeliveryPolicies = new DelayedDeliveryPolicies(2000, true);
-        admin.topics().setDelayedDeliveryPolicy(topicName, delayedDeliveryPolicies);
-        //wait for update
-        for (int i = 0; i < 50; i++) {
-            Thread.sleep(100);
-            if (admin.topics().getDelayedDeliveryPolicy(topicName) != null) {
-                break;
-            }
-        }
-        //2 Setup consumer and producer
-        @Cleanup
-        Consumer<String> consumer = pulsarClient.newConsumer(Schema.STRING)
-                .topic(topicName)
-                .subscriptionName("test-sub" + System.currentTimeMillis())
-                .subscriptionType(SubscriptionType.Shared)
-                .subscribe();
-
-        @Cleanup
-        Producer<String> producer = pulsarClient.newProducer(Schema.STRING)
-                .topic(topicName).create();
-        //3 Send delay message
-        for (int i = 0; i < 10; i++) {
-            producer.newMessage()
-                    .value("delayed-msg-" + i)
-                    .deliverAfter(4, TimeUnit.SECONDS)
-                    .sendAsync();
-        }
-        producer.flush();
-
-        //4 There will be no message in the first 3 seconds
-        assertNull(consumer.receive(3, TimeUnit.SECONDS));
-
-        Set<String> delayedMessages = new HashSet<>();
-        for (int i = 0; i < 10; i++) {
-            Message<String> msg = consumer.receive(4, TimeUnit.SECONDS);
-            delayedMessages.add(msg.getValue());
-            consumer.acknowledge(msg);
-        }
-        for (int i = 0; i < 10; i++) {
-            assertTrue(delayedMessages.contains("delayed-msg-" + i));
-        }
-        //5 Disable delayed delivery
-        delayedDeliveryPolicies.setActive(false);
-        admin.topics().setDelayedDeliveryPolicy(topicName, delayedDeliveryPolicies);
-        //wait for update
-        for (int i = 0; i < 50; i++) {
-            Thread.sleep(100);
-            if (!admin.topics().getDelayedDeliveryPolicy(topicName).isActive()) {
-                break;
-            }
-        }
-        producer.newMessage().value("disabled-msg").deliverAfter(5, TimeUnit.SECONDS).send();
-        //6 Delay deliver is disabled, so we can receive message immediately
-        Message<String> msg = consumer.receive(1, TimeUnit.SECONDS);
-        assertNotNull(msg);
-        consumer.acknowledge(msg);
-        //7 Set a very long tick time, so that trackDelayedDelivery will fail. we can receive msg immediately.
-        delayedDeliveryPolicies.setActive(true);
-        delayedDeliveryPolicies.setTickTime(Integer.MAX_VALUE);
-        admin.topics().setDelayedDeliveryPolicy(topicName, delayedDeliveryPolicies);
-        //wait for update
-        for (int i = 0; i < 50; i++) {
-            Thread.sleep(100);
-            if (admin.topics().getDelayedDeliveryPolicy(topicName).isActive()) {
-                break;
-            }
-        }
-        producer.newMessage().value("long-tick-msg").deliverAfter(5, TimeUnit.SECONDS).send();
-        msg = consumer.receive(1, TimeUnit.SECONDS);
-        assertNotNull(msg);
     }
 
     @Test
